+++ app/javascript/vue/tasks/otu/browse/main.js
@@ -1,10 +1,9 @@
 import { createApp } from 'vue'
+import { newStore } from './store/store'
 import App from './app.vue'
 import hotkey from 'plugins/v-hotkey'
 import HelpSystem from 'plugins/help/help'
 import en from './lang/en.js'
-
-import { newStore } from './store/store'
 
 function init () {
   const app = createApp(App)
@@ -18,7 +17,7 @@ function init () {
   app.mount('#vue-task-otu-browse')
 }
 
-$(document).on('turbolinks:load', () => {
+document.addEventListener('turbolinks:load', () => {
   if (document.querySelector('#vue-task-otu-browse')) {
     init()
   }